build-sys: only build col on glibc
col expects an EILSEQ failure of getwchar() to leave the undecoded data
in the buffer to be read by a subsequent getchar().

This is not guaranteed by the standard and for example musl libc does
not conform to this expectation.

As col is deprecated anyways just disable it on non-glibc instead of fixing it.

Signed-off-by: Thomas Weißschuh <thomas@t-8ch.de>