treewide: explicitly mark unused arguments
The autotools build used -Wno-unused-parameter to silence these warnings
for a few files. On meson however this configuration was not duplicated
leading to persistent warnings, preventing the usage of -Werror.

Instead of having to maintain the exceptions in two buildsystems, mark
the exceptions directly in the source code.

Afterward clean up autotools to not use -Wno-unused-parameter anymore.

Signed-off-by: Thomas Weißschuh <thomas@t-8ch.de>