util-linux/sys-utils test dmesg support for additional human-readable timestamp
This commit adds tests for new functionality in dmesg to support multiple timestamp formats.
The user can specify multiple timestamp formats with --time-format option and dmesg prints all the logs with desired timestamp formats in the sequence of user-provided formats.

This test verifies that output contains multiple formats in the sequence of input and it also tests for the special case where the user specifies delta format along with either ctime or raw.

Signed-off-by: Rishabh Thukral <rthukral@arista.com>